+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added `capacityset` package
+- Changed `batchDeleter` to leverage `capacityset`
 
 ## [v0.9.0]
 - Changed batchDeleter to expect the deathdate to be in Unix Nanoseconds